% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{DEBRA}
\alias{DEBRA}
\title{DEBRA - DESeq-based Barcode Representation Analysis}
\usage{
DEBRA(counts, control_names, condition_names, alpha = -Inf,
  method = "DESeq", trended = T, shrunkLFC = F, modified = T)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
DEBRA - DESeq-based Barcode Representation Analysis
}
\examples{
cat_function()
}
\keyword{cats}
