% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{DEBRADataSet}
\alias{DEBRADataSet}
\title{Construct a DEBRADataSet}
\usage{
DEBRADataSet(counts, control_names, condition_names, method = c("DESeq",
  "DESeq2(Wald)", "DESeq2(LRT)"), beta = -Inf, trended = T,
  shrunkLFC = F, modified = T, default_beta = 10)
}
\arguments{
\item{counts}{a data frame of non-negative read counts with columns of samples and rownames of barcode IDs; samples not included into analysis are allowed}

\item{control_names}{a character vector specifying the control samples (colnames of the counts data frame)}

\item{condition_names}{a character vector specifying the condition samples (colnames of the counts data frame)}

\item{method}{a character specifying the method used for inferring differentially represented barcodes}

\item{beta}{a numeric specifying beta value providing a lower read count threshold value for an independent filtering step; if beta = -Inf (default), the beta will be estimated from the read counts of condition samples}

\item{trended}{a logical specifying if the trended dispersion estimates should be used; if trended=FALSE, the shrunken dispersion estimates (as estimated by DESeq2) are used}

\item{modified}{a logical, if modified = F then the non-modified version of the correspondig method (DESeq, DESeq2(Wald) or DESeq2(LRT)) will be run; note independent filtering using beta threshold can still be applied}

\item{default_beta}{a numeric specifying the beta value used if the beta estimation is failed}

\item{shrinkLFC}{a logical specifying if the logFC values should be shrunken using "apeglm" shrinkage estimator}
}
\description{
Construct a DEBRADataSet
}
