% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{estimateBeta}
\alias{estimateBeta}
\title{Estimates the beta threshold for DEBRADataSet}
\usage{
estimateBeta(drb, s = 500, obs_window = 30, max_window = 150,
  KS_test_window = 30, default_beta = 10)
}
\arguments{
\item{drb}{DEBRADataSet object}

\item{s}{an integer specifying number of replicates for local Kolmogorov-Smirnov statistic estimation. Default is 1000}

\item{obs_window}{an integer specifying number of observations sampled from ordered mean counts values. Mean counts are calculated from condition samples}

\item{max_window}{a numeric specifying maximum size of the sampling window in read counts max(sampled_counts)-min(sampled_counts) < max_window}

\item{KS_test_window}{an integer specifying number of observations sampled from theoretical and empirical Kolmogorov-Smirnov statistics}

\item{default_beta}{a numeric specifying beta value used if the estimation has failed}
}
\description{
Estimates the beta threshold for DEBRADataSet
}
\examples{
drb<-estimateBeta(drb)
}
