% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filterPlot}
\alias{filterPlot}
\title{Visulaize independet filtering results}
\usage{
filterPlot(drb)
}
\arguments{
\item{drb}{DEBRADataSet object}
}
\description{
Visualize optimization of the number of rejections over the quantiles of the mean counts.
}
\examples{
filterPlot(drb)
}
