% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{independentFilteringDRB}
\alias{independentFilteringDRB}
\title{Independent filtering}
\usage{
independentFilteringDRB(drb, filter_FDR = NULL, beta = NULL)
}
\arguments{
\item{drb}{DEBRADataSet object}

\item{filter_FDR}{user-specified FDR threshold for calculating number of null hypothesis rejections}

\item{beta}{beta threshold provides a lower read count value for an independent filtering step; if not provided, beta value is extracted from DEBRADataSet}
}
\description{
Independent filtering step for DEBRA analysis results
}
\examples{
drb<-independentFilteringDRB(drb)
}
