% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{KS_plot}
\alias{KS_plot}
\title{Plot theoretical and empirical Kolmogorov-Smirnov (KS) statistic values}
\usage{
KS_plot(drb)
}
\arguments{
\item{drb}{DEBRADataSet object}
}
\description{
Plot local estimates of theoretical and empirical Kolmogorov-Smirnov test statistic values used for beta threshold estimation
}
\examples{
KS_plot(drb)
}
