% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{overlapFit}
\alias{overlapFit}
\title{Plot mean count versus overlap values between theoretical and empirical Kolmogorov-Smirnov (KS) test statistics and the sigmoid fit}
\usage{
overlapFit(drb)
}
\arguments{
\item{drb}{DEBRADataSet object}
}
\description{
This function plots the local overlap between Gamma-modelled empirical and theoretical Kolmogorov-Smirnov statistic values fitted with four parameters sigmoid function.
}
\examples{
overlapFit(drb)
}
