% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{resultsDRB}
\alias{resultsDRB}
\title{Extracts a result table from DEBRADataSet object}
\usage{
resultsDRB(drb, filtered = T, sort = T)
}
\arguments{
\item{drb}{DEBRADataSet object}

\item{filtered}{logical specifying if filtered results should be extracted; default is TRUE}
}
\description{
resultsDRB extracts a DEBRA analysis results from a DEBRADataSet object.
}
\examples{
resultsDRB(drb, filtered = T)
}
