% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{testDRB}
\alias{testDRB}
\title{Statistical significane test for differential barcode representation using parameters provided in DEBRADataSet}
\usage{
testDRB(drb)
}
\arguments{
\item{drb}{DEBRADataSet}
}
\description{
Statistical significane test for differential barcode representation using parameters provided in DEBRADataSet
}
\examples{
drb<-testDRB(drb)
}
